+++ index.php
@@ -1,8 +1,13 @@
 <?php
+/**
+ * @deprecated 0.2:1.0 Use of package root as web root is deprecated in favor of `public/` directory.
+ */
 
 use Rarst\ReleaseBelt\Application;
 
-require __DIR__ . '/vendor/autoload.php';
+trigger_error('Use of package root as web root is deprecated in favor of `public/` directory.', E_USER_DEPRECATED);
+
+require __DIR__.'/vendor/autoload.php';
 
 $configPath = __DIR__.'/config.php';